Move plugin machinery into dedicated module in line with pyproject.toml
